requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
curl-cffi==0.12.0
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.17.0
duckduckgo-search==8.1.1
email-validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
frozendict==2.4.6
gitdb==4.0.12
gitpython==3.1.45
greenlet==3.2.3
groq==0.30.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
lxml==6.0.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
multitasking==0.0.12
numpy==2.0.2
packaging==25.0
pandas==2.3.1
peewee==3.18.2
phidata==2.7.10
platformdirs==4.3.8
primp==0.15.0
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
requests==2.32.4
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
sentry-sdk==2.33.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
starlette==0.47.2
tomli==2.2.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yfinance==0.2.65
